requirements.txt
requests
pymongo
python-dotenv
boto3
pulp

